Delete All comments, Change SideNav Icons Set UI ko CLient
