+++ package.json
@@ -23,10 +23,9 @@
     "@fontsource/plus-jakarta-sans": "^5.1.0",
     "@fontsource/roboto-mono": "^5.1.0",
     "@hookform/resolvers": "3.9.1",
-    "@mui/icons-material": "^5.11.0",
+    "@mui/icons-material": "^5.16.8",
     "@mui/lab": "^5.0.0-alpha.170",
-
-    "@mui/material": "^5.11.0",
+    "@mui/material": "^5.16.8",
     "@mui/styles": "^6.1.9",
     "@mui/system": "^6.1.9",
     "@mui/utils": "^6.1.9",
@@ -36,6 +35,7 @@
     "apexcharts": "3.49.2",
     "autoprefixer": "^10.4.20",
     "axios": "^1.7.7",
+    "clsx": "^2.1.1",
     "css-loader": "^7.1.2",
     "dayjs": "1.11.11",
     "moment": "^2.30.1",
