+++ postcss.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   plugins: {
-    autoprefixer: {}, // Automatically adds vendor prefixes for CSS
+    autoprefixer: {}, 
   },
 };
