--- src/components/core/logo.tsx
@@ -1,56 +1,0 @@
-// 'use client';
-
-// import * as React from 'react';
-// import Box from '@mui/material/Box';
-// import { useColorScheme } from '@mui/material/styles';
-
-// import { NoSsr } from '@/components/core/no-ssr';
-
-// const HEIGHT = 120; // Increased default height
-// const WIDTH = 120;  // Increased default width
-
-// type Color = 'dark' | 'light';
-
-// export interface LogoProps {
-//   color?: Color;
-//   emblem?: boolean;
-//   height?: number;
-//   width?: number;
-// }
-
-// export function Logo({ color = 'dark', emblem, height = HEIGHT, width = WIDTH }: LogoProps): React.JSX.Element {
-//   let url: string;
-
-//   if (emblem) {
-//     url = color === 'light' ? '/assets/vehware.png' : '/assets/vehware.png';
-//   } else {
-//     url = color === 'light' ? '/assets/vehware.png' : '/assets/vehware.png';
-//   }
-
-//   return <Box alt="logo" component="img" height={height} src={url} width={width} />;
-// }
-
-// export interface DynamicLogoProps {
-//   colorDark?: Color;
-//   colorLight?: Color;
-//   emblem?: boolean;
-//   height?: number;
-//   width?: number;
-// }
-
-// export function DynamicLogo({
-//   colorDark = 'light',
-//   colorLight = 'dark',
-//   height = HEIGHT,
-//   width = WIDTH,
-//   ...props
-// }: DynamicLogoProps): React.JSX.Element {
-//   const { colorScheme } = useColorScheme();
-//   const color = colorScheme === 'dark' ? colorDark : colorLight;
-
-//   return (
-//     <NoSsr fallback={<Box sx={{ height: `${height}px`, width: `${width}px` }} />}>
-//       <Logo color={color} height={height} width={width} {...props} />
-//     </NoSsr>
-//   );
-// }
