+++ src/contexts/user-context.tsx
@@ -46,9 +46,7 @@ export function UserProvider({ children }: UserProviderProps): React.JSX.Element
   React.useEffect(() => {
     checkSession().catch((err: unknown) => {
       logger.error(err);
-      // noop
     });
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- Expected
   }, []);
 
   return <UserContext.Provider value={{ ...state, checkSession }}>{children}</UserContext.Provider>;
