+++ src/lib/logger.ts
@@ -1,6 +1,4 @@
-/* eslint-disable no-console -- Allow */
 
-// NOTE: A tracking system such as Sentry should replace the console
 
 export const LogLevel = { NONE: 'NONE', ERROR: 'ERROR', WARN: 'WARN', DEBUG: 'DEBUG', ALL: 'ALL' } as const;
 
@@ -63,8 +61,7 @@ export class Logger {
   }
 }
 
-// This can be extended to create context specific logger (Server Action, Router Handler, etc.)
-// to add context information (IP, User-Agent, timestamp, etc.)
+
 
 export function createLogger({ prefix, level }: LoggerOptions = {}): Logger {
   return new Logger({ prefix, level });
