+++ src/pages/document.tsx
@@ -1,4 +1,3 @@
-// src/pages/_document.tsx
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 
 class MyDocument extends Document {
@@ -6,7 +5,7 @@ class MyDocument extends Document {
     return (
       <Html lang="en">
         <Head>
-          {/* You can add custom meta tags, styles, or scripts here */}
+       
         </Head>
         <body>
           <Main />
