+++ src/types/nav.d.ts
@@ -7,9 +7,6 @@ export interface NavItemConfig {
   icon?: string;
   href?: string;
   items?: NavItemConfig[];
-  // Matcher cannot be a function in order
-  // to be able to use it on the server.
-  // If you need to match multiple paths,
-  // can extend it to accept multiple matchers.
+
   matcher?: { type: 'startsWith' | 'equals'; href: string };
 }
